Add truecolor support for Windows, if available
- Update to latest tcell which has 24 bit Windows support
- light renderer under Windows defaults to Dark256, if possible
